Refactor v2 validator conditions
